+++ HISTORY.rst
@@ -3,6 +3,17 @@
 History
 -------
 
+0.6.1 (2013-11-23)
+~~~~~~~~~~~~~~~~~~
+
+- Bug fixing: --model-fields option with absolute field names was not
+  compatible with multi-label classification models.
+- Changing resource type checking function.
+- Bug fixing: evaluations did not use the given combination method.
+- Bug fixing: evaluation of an ensemble had turned into evaluations of its
+              models.
+- Adding pruning to the ensemble creation configuration options
+
 0.6.0 (2013-11-08)
 ~~~~~~~~~~~~~~~~~~